refactor: move the hook logic to a new file
